comparison sign mismatch fixes
git-svn-id: https://nocs.stanford.edu/cgi-bin/svn.cgi/booksim2.0/branches/classes@4598 ce7aba0d-9f81-40a6-a495-cecae84b0515

+++ src/workload.cpp
@@ -121,7 +121,7 @@ void Workload::advanceTime()
     int const source = _deferred_nodes.front();
     _deferred_nodes.pop();
     _pending_nodes.push(source);
-    assert(_pending_nodes.size() <= _nodes);
+    assert(_pending_nodes.size() <= (size_t)_nodes);
   }
 }
 
@@ -143,7 +143,7 @@ void Workload::defer()
   assert(!_pending_nodes.empty());
   int const source = _pending_nodes.front();
   _deferred_nodes.push(source);
-  assert(_deferred_nodes.size() <= _nodes);
+  assert(_deferred_nodes.size() <= (size_t)_nodes);
   _pending_nodes.pop();
 }
 
@@ -184,10 +184,10 @@ void SyntheticWorkload::reset()
   for(int source = 0; source < _nodes; ++source) {
     if(_injection->test(source)) {
       _pending_nodes.push(source);
-      assert(_pending_nodes.size() <= _nodes);
+      assert(_pending_nodes.size() <= (size_t)_nodes);
     } else {
       _sleeping_nodes.push(source);
-      assert(_sleeping_nodes.size() <= _nodes);
+      assert(_sleeping_nodes.size() <= (size_t)_nodes);
     }
   }
 }
@@ -208,10 +208,10 @@ void SyntheticWorkload::advanceTime()
     }
     if(generated) {
       _pending_nodes.push(source);
-      assert(_pending_nodes.size() <= _nodes);
+      assert(_pending_nodes.size() <= (size_t)_nodes);
     } else {
       _sleeping_nodes.push(source);
-      assert(_sleeping_nodes.size() <= _nodes);
+      assert(_sleeping_nodes.size() <= (size_t)_nodes);
     }
   }
 }
@@ -262,7 +262,7 @@ void SyntheticWorkload::inject(int pid)
   int const source = _pending_nodes.front();
   _pending_nodes.pop();
   _sleeping_nodes.push(source);
-  assert(_sleeping_nodes.size() <= _nodes);
+  assert(_sleeping_nodes.size() <= (size_t)_nodes);
 }
 
 TraceWorkload::TraceWorkload(int nodes, string const & filename, 
@@ -307,7 +307,7 @@ void TraceWorkload::_refill(int time)
       if(((_scale > 0) ? (time / _scale) : (time * -_scale)) <= _time) {
 	if(_ready_packets[source].empty()) {
 	  _pending_nodes.push(source);
-	  assert(_pending_nodes.size() <= _nodes);
+	  assert(_pending_nodes.size() <= (size_t)_nodes);
 	}
 	_ready_packets[source].push(_next_packet);
 	_next_source = -1;
@@ -346,7 +346,7 @@ void TraceWorkload::advanceTime()
     if(((_scale > 0) ? (time / _scale) : (time * -_scale)) <= _time) {
       if(_ready_packets[_next_source].empty()) {
 	_pending_nodes.push(_next_source);
-	assert(_pending_nodes.size() <= _nodes);
+	assert(_pending_nodes.size() <= (size_t)_nodes);
       }
       _ready_packets[_next_source].push(_next_packet);
       _next_source = -1;
@@ -409,6 +409,6 @@ void TraceWorkload::inject(int pid)
   _ready_packets[source].pop();
   if(!_ready_packets[source].empty()) {
     _deferred_nodes.push(source);
-    assert(_deferred_nodes.size() <= _nodes);
+    assert(_deferred_nodes.size() <= (size_t)_nodes);
   }
 }
